created different pages
