+++ lib/main.dart
@@ -10,6 +10,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       title: 'Present Mam',
       theme: ThemeData(
         // This is the theme of your application.
