+++ lib/pages/classes_page.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class Classes extends StatelessWidget {
+  const Classes({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      "Classes"
+    );
+  }
+}
