+++ lib/pages/subjects_dart.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+ class Subjects extends StatelessWidget {
+   const Subjects({Key? key}) : super(key: key);
+
+   @override
+   Widget build(BuildContext context) {
+     return Text("Subjects");
+   }
+ }
